TubeFetch.api/requirements.txt
spotipy==2.25.0
python-dotenv==1.0.1
youtube-search==2.1.2
Werkzeug==3.1.3
yt-dlp==2025.1.26
ffmpeg==1.4
joblib==1.4.2
joblib==1.4.2
Flask==3.1.0
flask-cors==5.0.0